+++ Pandora/Paradigm/Basis/Constant.hs
@@ -3,6 +3,7 @@ module Pandora.Paradigm.Basis.Constant (Constant (..)) where
 import Pandora.Core.Morphism (($))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
+import Pandora.Pattern.Functor.Invariant (Invariant (invmap))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
 
@@ -13,6 +14,9 @@ instance Covariant (Constant a) where
 
 instance Contravariant (Constant a) where
 	_ >$< Constant x = Constant x
+
+instance Invariant (Constant a) where
+	invmap _ _ (Constant x) = Constant x
 
 instance Traversable (Constant a) where
 	Constant x ->> _ = point (Constant x)