+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 * Replace `flip` on `?` type operator
 * Remove type parameter from `Stack` type synonymous
 * Split `structure` and `value` in `Nonempty` type family
+* Define `Invariant` instance for `Constant` datatype
+* Extract `Capture` from `Observable` into separated module